refacto - the methodology  blueprint is now embedded in a class. This is motivated by translation that cannot work otherwise. We apply this refactorization to this branch to ease future merge
