+++ pages/ai.py
@@ -21,11 +21,10 @@ from dash_extensions.enrich import DashBlueprint, html
 from blueprints.form.form_blueprint import FormBlueprint
 from blueprints.import_export.import_export_blueprint import ImportExportBlueprint
 from blueprints.metrics.metrics_blueprint import get_metrics_blueprint
-from blueprints.methodology.methodology_blueprint import get_methodology_blueprint
+from blueprints.methodology.methodology_blueprint import MethodologyBlueprint
 
 import blueprints.metrics.metrics_layout as metrics_layout
 import blueprints.metrics.utils as metrics_utils
-import blueprints.methodology.methodology_layout as methodo_layout
 
 from utils.graphics import MY_COLORS
 from utils.handle_inputs import get_available_versions, filter_wrong_inputs, clean_non_used_inputs_for_export,  open_input_csv_and_comment, read_base_form_inputs_from_csv, AI_PAGE_DEFAULT_VALUES, validate_ai_page_specific_inputs
@@ -66,7 +65,7 @@ inference_form = FormBlueprint(
 # 2000 milliseconds to avoid rendering bugs of the server fields
 import_export = ImportExportBlueprint(id_prefix=AI_PAGE_ID_PREFIX, csv_flushing_delay=2500) 
 
-methodo_content = get_methodology_blueprint(
+methodo_content = MethodologyBlueprint(
     id_prefix=AI_PAGE_ID_PREFIX,
     additional_formula_content=dcc.Markdown(
         '''
@@ -104,6 +103,172 @@ appVersions_options = get_available_versions()
 
 ###################################################
 # DEFINE APP LAYOUT
+
+def get_training_help_content(title: str):
+    '''
+    This layout is designed to be methodological content located in the
+    'Help' tab of the training form in the AI page. Its purpose is to help the
+    user with the training form requirements.
+    '''
+    return html.Div(
+        [
+            html.H3(title),
+            
+            html.Div(
+                [
+
+                    html.H4('Overall description'),
+
+                    dcc.Markdown(
+                        '''
+                        The training phase of your AI system includes different stages:
+                        R&D experiments, the final training of your model, and potential retraining runs 
+                        (more details are given below).
+                        
+                        We invite you to include all these stages in this form. 
+                        __Start by filling-in the form based on the main training run.
+                        Then use the input fields available at the bottom of the form to 
+                        estimate the impact of R&D training and/or retraining__.
+                        Some R&D use cases are illustrated in more details below. 
+
+                        This approach may not work for all cases (e.g. if retraining is done with different
+                        hardware) but it offers a reasonable estimate. 
+                        Besides, there is always the option of rerunning the calculator separately.
+                        ''',
+                        className='form-help-markdown'
+                    ),
+                ],
+                className='form-help-subsection',
+            ),
+
+            html.Div(
+                [
+                    html.Hr(),
+                ],
+                className='Hr_div'
+            ),
+
+            html.Div(
+                [
+                    html.H4('Definitions and practical tips'),
+
+                    dcc.Markdown(
+
+                        ''' 
+                        __Main/final training stage__: the computations performed to achieve the final model 
+                        of your AI solution. 
+                        It can either correspond to training from scratch a custom model 
+                        or fine-tuning an existing model.
+
+                        __R&D training__: the compute involved in the research and development phase before the 
+                        final run (e.g. hyper-parameters search).
+                        This is included by added R&D as a fraction of the training time, 
+                        e.g. enter 2 if you assume that in total R&D represents 
+                        double the compute resources of the final run. 
+                        There is no typical value for that, from a small fraction in case of 
+                        well defined straightforward models to hundreds in more complex models requiring 
+                        extensive searches. For example, when studying a 176 billon parameters LLM,
+                        Luccioni et al. \[1\] estimated that intermediate models training and evaluation accounted 
+                        for approximately 150% of their main training consumption.
+                        
+                        __Retraining__: any additional training runs performed after the deployment of your AI system.
+                        For consistent reporting, you are invited to take into account all retraining happening 
+                        over your reporting period.
+                        '''
+                    ),
+
+                    dcc.Markdown(
+                        '\[1\] A. S. Luccioni, S. Viguier, and A.-L. Ligozat, '
+                        '“Estimating the Carbon Footprint of BLOOM, a 176B Parameter Language Model,” Journal '
+                        'of Machine Learning Research, vol. 24, no. 253, pp. 1–15, 2023',
+                        className='footnote citation-report',
+                        style={'margin-top': '8px'}
+                    ),
+                ],
+                className='form-help-subsection',
+            ),
+        ],
+        className='form-help-container pretty-container container'
+    )
+
+
+def get_inference_help_content(title: str):
+    '''
+    This layout is designed to be methodological content located in the
+    'Help' tab of the inference form in the AI page. Its purpose is to help the
+    user with the inference form requirements.
+    '''
+    return html.Div(
+        [
+            html.H3(title),
+
+            html.Div(
+                [
+                    html.H4('Overall description'),
+
+                    dcc.Markdown(
+                        '''
+                        This is to quantify the environmental impacts of the inference phase of your AI system.
+                        __We distinguish between two types of inference: block, or one-shot, inference 
+                        (you make predictions once and for all)  and continuous inference 
+                        (the model makes predictions continuously over time, e.g. a chatbot)__.
+
+                        By default, the form is in 'block inference' mode but you can activate 
+                        the continuous mode using the switch at the top.
+                        '''
+                    )
+                ],
+                className='form-help-subsection',
+            ),
+
+            html.Div(
+                [
+                    html.Hr(),
+                ],
+                className='Hr_div',
+            ),
+
+            html.Div(
+                [
+                    html.H4('Definitions and practical tips'),
+
+                    dcc.Markdown(
+                        '''
+                        __Block (one-shot) inference__: for a system where prediction is made on a one-off basis 
+                        (or repeated occasionally).
+                        It may be used to process a data set as a whole or to build one-day or one-week strategy. 
+                        If multiple block inferences happen over your reporting period, 
+                        we invite you to quantify the resource needs for one inference block 
+                        and then to use the multiplicative factor.
+                        ''',
+                        style={'margin-bottom': '6px'}
+                    ),
+
+                    dcc.Markdown(
+                        '''
+                        __Continuous inference__: corresponds to an AI service that is requested on demand 
+                        by users or other software systems (e.g. chatbot). 
+                        This inference workload does not follow a strict scheduling, making it harder to quantify. 
+                        In this mode, we invite you to estimate the resource usage 
+                        over a period of time of your choice, the so-called "input data time span”. 
+                        The results are then scaled up over the total reporting period. 
+                        For instance, if choosing a reporting scope of 1 year and filling 
+                        the form in continuous inference mode with an `input data time span` of 1 month,
+                        then your environmental impacts correspond to the monthly results multiplied by 12.
+                        
+                        It is worth keeping in mind that __the reporting period only impacts the results
+                        in the 'continuous inference' situation__.
+                        ''',
+                        # style={'margin-bottom': '12px'}
+                    ),
+
+                ],
+                className='form-help-subsection',
+            )
+        ],
+        className='form-help-container container'
+    )
+
 
 def get_ai_page_layout():
     page_layout = html.Div(
@@ -222,7 +387,7 @@ def get_ai_page_layout():
                                 ]
                             ),
                             dmc.TabsPanel(children=training_form.embed(AI_PAGE), value='form'),
-                            dmc.TabsPanel(children=methodo_layout.get_training_help_content(''), value='help'),
+                            dmc.TabsPanel(children=get_training_help_content(''), value='help'),
                         ],
                         value="form",
                         className='tab-container',
@@ -256,7 +421,7 @@ def get_ai_page_layout():
                                 ]
                             ),
                             dmc.TabsPanel(children=inference_form.embed(AI_PAGE), value='form'),
-                            dmc.TabsPanel(children=methodo_layout.get_inference_help_content(''), value='help'),
+                            dmc.TabsPanel(children=get_inference_help_content(''), value='help'),
                         ],
                         value="form",
                         className='tab-container',
